Undo `ValueConversions` renaming
